+++ xarray/tests/test_coding_times.py
@@ -26,7 +26,7 @@ from xarray.coding.times import (
 from xarray.coding.variables import SerializationWarning
 from xarray.conventions import _update_bounds_attributes, cf_encoder
 from xarray.core.common import contains_cftime_datetimes
-from xarray.testing import assert_equal
+from xarray.testing import assert_equal, assert_identical
 
 from . import (
     arm_xfail,
@@ -1049,3 +1049,21 @@ def test__encode_datetime_with_cftime():
     expected = cftime.date2num(times, encoding_units, calendar)
     result = _encode_datetime_with_cftime(times, encoding_units, calendar)
     np.testing.assert_equal(result, expected)
+
+
+@pytest.mark.parametrize("calendar", ["gregorian", "Gregorian", "GREGORIAN"])
+def test_decode_encode_roundtrip_with_non_lowercase_letters(calendar):
+    # See GH 5093.  This is primarily relevant for standard calendar times when
+    # cftime is not installed (cftime.num2date already lowers the calendar name
+    # internally).
+    times = [0, 1]
+    units = "days since 2000-01-01"
+    attrs = {"calendar": calendar, "units": units}
+    variable = Variable(["time"], times, attrs)
+    decoded = conventions.decode_cf_variable("time", variable)
+    encoded = conventions.encode_cf_variable(decoded)
+
+    # Use assert_identical to ensure that the calendar attribute maintained its
+    # original form throughout the roundtripping process, uppercase letters and
+    # all.
+    assert_identical(variable, encoded)